update presence.py delete pipe.py
